tried to let MarktplaatsIT work Changed settings in persistence.xml in test/resources added dummydata for demo changed boddao get by id from int to long

+++ src/main/java/nl/corne/marktplaats/model/advertentie/AdvertentieDAOInterface.java
@@ -6,7 +6,7 @@ import nl.corne.marktplaats.model.gebruiker.Gebruiker;
 import java.util.List;
 
 public interface AdvertentieDAOInterface extends DAOInterface<Advertentie> {
-    Advertentie get(int id);
+    Advertentie get(long id);
 
     List<Advertentie> getAllAdvertentieFromGebruiker(Gebruiker gebruiker);
 }
